id: generate id with hasing
The current sets of hash functions are very limited, collitions are likely with this hash, but there is still the other option of generating an incemental id or extending the P4 with better hash algorithm. Collition detection could be done, but hard and out of the scope of this work.